Resolved comments: error class format, class hierarchy
